requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-google-genai
sentence-transformers
chromadb
pypdf
python-dotenv
tiktoken
tenacity
pymupdf